requirements.txt
pika>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
tenacity>=8.2.0